+++ attribution/src/main/java/com/affise/attribution/init/AffiseInitProperties.kt
@@ -2,6 +2,7 @@ package com.affise.attribution.init
 
 import com.affise.attribution.events.autoCatchingClick.AutoCatchingType
 import com.affise.attribution.network.CloudConfig
+import com.affise.attribution.settings.AffiseConfig
 import com.affise.attribution.settings.OnInitErrorHandler
 import com.affise.attribution.settings.OnInitSuccessHandler
 
@@ -20,6 +21,7 @@ data class AffiseInitProperties @JvmOverloads constructor(
     val domain: String? = CloudConfig.defaultDomain,
     val onInitSuccessHandler: OnInitSuccessHandler? = null,
     val onInitErrorHandler: OnInitErrorHandler? = null,
+    val configValues: Map<AffiseConfig, Any> = emptyMap()
 ) {
 
     init {