+++ module-meta/src/main/java/com/affise/attribution/module/meta/usecase/MetaReferrerUseCase.kt
@@ -1,0 +1,6 @@
+package com.affise.attribution.module.meta.usecase
+
+interface MetaReferrerUseCase {
+
+    fun metaReferrer(): Map<String, Any>
+}